requirements.txt
feast==0.34.1
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
redis==5.0.1
pandas==2.1.3
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
apache-flink==1.17.1
pyarrow==14.0.1
xgboost==2.0.2
imbalanced-learn==0.11.0
scikit-learn==1.3.2
lightgbm==4.1.0
catboost==1.2.2
optuna==3.4.0
shap==0.43.0
plotly==5.17.0
aiohttp==3.9.1
docker==6.1.3
psutil==5.9.6
networkx==3.2.1
scipy==1.11.4
tensorflow==2.15.0
keras==2.15.0
lime==0.2.0.1
streamlit==1.28.2
sqlite3
joblib==1.3.2
